+++ src/ByteDev.Common/NameValueCollectionExtensions.cs
@@ -23,10 +23,8 @@ namespace ByteDev.Common
                 throw new ArgumentNullException(nameof(source));
 
             if (source.Get(key) == null)
-            {
                 return source.AllKeys.Contains(key);
-            }
-
+ 
             return true;
         }
     }